MR: 네이버 검색 상품 DB에 즐겨찾기(로그인한 유저의 userId 저장, CSRF 임시로 비활성화)

+++ src/main/java/com/mysite/sbbfinal/mapper/UserMapper.java
@@ -16,7 +16,7 @@ public interface UserMapper {
     List<User> findAll();  
     List<Role> findRolesByUserId(@Param("userId") Long userId);  
     User findById(@Param("id") Long id); 
-    int getUserIdByUsername(String username);
+    Long getUserIdByUsername(String username);
     
     List<Role> getAllRoles();
     void deleteUserRole(@Param("userId") Long userId, @Param("roleId") Long roleId);
